Consolidate preexisting form526 swagger docs under foldermodule
