+++ app/controllers/v0/apidocs_controller.rb
@@ -159,7 +159,7 @@ module V0
       Swagger::Schemas::BB::HealthRecords,
       Swagger::Schemas::Countries,
       Swagger::Schemas::ConnectedApplications,
-      Swagger::Schemas::RatedDisabilities,
+      Swagger::Schemas::Form526::RatedDisabilities,
       Swagger::Schemas::Email,
       Swagger::Schemas::Errors,
       Swagger::Schemas::EVSSAuthError,
@@ -170,6 +170,10 @@ module V0
       Swagger::Schemas::Form526::Form4142,
       Swagger::Schemas::Form526::Form526SubmitV2,
       Swagger::Schemas::Form526::Form8940,
+      Swagger::Schemas::Form526::JobStatus,
+      Swagger::Schemas::Form526::RatedDisabilities,
+      Swagger::Schemas::Form526::SubmitDisabilityForm,
+      Swagger::Schemas::Form526::SuggestedConditions,
       Swagger::Schemas::Gibct::CalculatorConstants,
       Swagger::Schemas::Gibct::Institutions,
       Swagger::Schemas::Health::Folders,
@@ -187,9 +191,6 @@ module V0
       Swagger::Schemas::PPIU,
       Swagger::Schemas::SavedForm,
       Swagger::Schemas::States,
-      Swagger::Schemas::SubmitDisabilityForm,
-      Swagger::Schemas::Form526JobStatus,
-      Swagger::Schemas::SuggestedConditions,
       Swagger::Schemas::TermsAndConditions,
       Swagger::Schemas::UploadSupportingEvidence,
       Swagger::Schemas::VAFacilities,
